us_visa/utils/__init__.py
import os
import sys
import numpy as np
import dill
import yaml
from pandas import DataFrame

from us_visa.exception import USVisaException
from us_visa.logger import logging


def read_yaml_file(file_path: str) -> dict:
    """Reads a YAML file and returns its contents as a dictionary."""
    try:
        with open(file_path, "r") as yaml_file:
            return yaml.safe_load(yaml_file)
    except Exception as e:
        raise USVisaException(e, sys) from e


def write_yaml_file(file_path: str, content: object, replace: bool = False) -> None:
    """Writes content to a YAML file."""
    try:
        if replace and os.path.exists(file_path):
            os.remove(file_path)

        os.makedirs(os.path.dirname(file_path), exist_ok=True)

        with open(file_path, "w") as file:
            yaml.dump(content, file)

    except Exception as e:
        raise USVisaException(e, sys) from e


def save_object(file_path: str, obj: object) -> None:
    """Saves a Python object using dill."""
    logging.info("Entered save_object()")
    try:
        os.makedirs(os.path.dirname(file_path), exist_ok=True)
        with open(file_path, "wb") as file_obj:
            dill.dump(obj, file_obj)
        logging.info("Exited save_object()")
    except Exception as e:
        raise USVisaException(e, sys) from e


def load_object(file_path: str) -> object:
    """Loads a Python object using dill."""
    logging.info("Entered load_object()")
    try:
        with open(file_path, "rb") as file_obj:
            obj = dill.load(file_obj)
        logging.info("Exited load_object()")
        return obj
    except Exception as e:
        raise USVisaException(e, sys) from e


def save_numpy_array_data(file_path: str, array: np.ndarray) -> None:
    """Saves a numpy array to a file."""
    try:
        os.makedirs(os.path.dirname(file_path), exist_ok=True)
        with open(file_path, "wb") as file_obj:
            np.save(file_obj, array)
    except Exception as e:
        raise USVisaException(e, sys) from e


def load_numpy_array_data(file_path: str) -> np.ndarray:
    """Loads a numpy array from a file."""
    try:
        with open(file_path, "rb") as file_obj:
            return np.load(file_obj, allow_pickle=True)
    except Exception as e:
        raise USVisaException(e, sys) from e


def drop_columns(df: DataFrame, cols: list) -> DataFrame:
    """Drops specified columns from a pandas DataFrame."""
    logging.info("Entered drop_columns()")
    try:
        df = df.drop(columns=cols)
        logging.info("Exited drop_columns()")
        return df
    except Exception as e:
        raise USVisaException(e, sys) from e
