requirements.txt
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost
pymongo
from_root
evidently==0.2.8
dill
pyYAML
neuro_mf
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
python-multipart
-e .


#  if you are getting mongoDB error use this smd


pip uninstall -y pymongo motor mongoengine django
pip install -u "pymongo>=4.7" dnspython certifi













